test(streamwall): cover the overlay error tile rendering
Extract the per-view tile content out of overlay.tsx into an exported
OverlayViewTile component, so the error-tile branch can be rendered
and asserted on directly instead of only being reachable through the
top-level Overlay/App tree (which self-renders to document.body on
import).

Adds happy-dom + a preact/compat-aware vitest config for the streamwall
package (mirroring streamwall-control-ui's setup) and component tests
asserting that errored views render the warning badge and reason,
non-error views render neither, and the heading falls back to
"Stream error" when there is no label or reason.
