+++ packages/streamwall/package.json
@@ -34,6 +34,7 @@
     "@types/ws": "^8.18.1",
     "@types/yargs": "^17.0.35",
     "electron": "^43.1.0",
+    "happy-dom": "^20.10.6",
     "ts-node": "^10.9.2",
     "typescript": "^6.0.3",
     "vite": "^8.1.3",
